[refactor]: Change client component to sever component(dashboard page)
